Draw the baize background using a ClutterTexture subclass.
A special subclass of ClutterTexture has been added which draws its
texture repeated to the full size of the stage it is in. This saves
the need to listen for resize events on the stage.

The expose handler for the board has been disabled so that it doesn't
paint over Clutter's drawing.

svn path=/trunk/; revision=8130