aisleriot/rules/valentine.scm: Be more strict about allowing drops
It used to be possible to move a card onto the waste or the stock, as
long as the ranks and suit are correct. This change prevents that.
Dropping a card from one slot to the same slot is also no longer
allowed, as that would create an empty undo entry.

svn path=/trunk/; revision=8679