[aisleriot-slot-render] Fix raise animation
The raise animation became broken in revision 8475 when the card was
not being flipped because the code to retrieve the card width was
moved so that it is only valid if the card is flipped. Now the card
size is always retrieved by querying the size of the actor instead of
retrieving the card texture directly.

svn path=/trunk/; revision=8806