[aisleriot-board] Check for animations in an idle handler
Instead of trying to check for animations whenever it thinks the
scheme engine might have changed something (such as after a button
press or drag finished) it now checks them in response to the slot
changed signal. This catches more circumstances than before so that
the deal button and undo buttons work properly. The actual check is
delayed to an idle handler to be sure that the check is only performed
once if the game changes multiple slots.

svn path=/trunk/; revision=8812