libgames-support: Add games_show_uri()
Implement a helper function to open a URI. On gtk+ >= 2.14, this just
uses gtk_show_uri(). Otherwise, on hildon, it uses osso to show the URI
in the browser, and otherwise uses xdg-open with a fallback to
gnome-open.
Supported URI schemes are file:, http[s]:. On the gnome platform, ghelp:
is supported as well.