aisleriot: Change licence to GPL 3+
Relicence the libgamescard library code from LGPL 2+ to LGPL 3+ (resp.
GPL 2+ to GPL 3+), and aisleriot from GPL 2+ to GPL 3+. The
libgamescards licence changes also implies a blackjack licence change
from GPL 2+ to GPL 3+.
Add COPYING.GPL3 containing the text of the GPL version 3, and
COPYING.README with a short summary of which code is licensed how.