aisleriot: maemo5: Redefine stock icons using a custom gtkrc file
Since the stock gtk icons we use aren't available on the device in the
required toolbar icon size (48x48), and to use icons that look better in
the device's theme, re-define the stock icons using a custom gtkrc file
on maemo5.

This is an initial try; there may be better choices for which hildon
icons to use for which stock icon.

Bug #584669.