aisleriot: Fix card theme menu
The order in which the card theme directories are scanned assumed
that the first theme with a given pref name is retained, but
the code actually replaced that one when further themes with the same
pref name were found. Fixing that fixes the card theme menu in
aisleriot to show the current theme correctly.