Define new hint functions and deprecate (get-name).
See bug 551859.

The inner workings of (hint-move) and (hint-click) may change in the future,
but its usage will not. The idea is that if we need to change it for i18n
reasons, or we want to show the hints visually instead of as text, we can do
that by modifying those functions instead of every game.