kansas: Add "move tableau to self" type hints.
These are cases where one can reveal a card in the tableau with the rank
equal to the top card's rank (or empty a tableau pile if the bottom card's
rank is immediately below the top card's rank) by first moving part of that
pile to another tableau, then moving from the same pile to the same pile
again.

This also removes hints that attempt to group cards of the same suit in the
tableau, as they are unnecessary and would combine with this new type of hint
to create situations where the hints suggest moving a stack endlessly between
two tableau piles.

For bug 660297.