guile: Install the compiled files not the sources
On guile 2.x, install the compiled .go files instead of the source .scm files.

Remove the check for guile-tools in configure, and instead check for guile.
guile-tools however aborts the compile because it doesn't know about our internal
(aisleriot interface) module. So we copy it into our repo (the licence is LGPL3+
which is compatible with ours), and add a dummy module definition to it.