games: Update src/get_titles.pl script for new game filenames
The guide for adding new games, games/README, instructs developers to
run src/get_titles.pl in order to update the list of game names in
src/game-names.h. However, since the games were renamed from "_" to "-"
in a4efa141 (guile: Canonicalise filenames, 2011-11-30), the script has
not been updated to handle the new filenames, and outputs like this:

  /* Translators: this string is the name of a game of patience.
   If there is an established standard name for this game in your
   locale, use that; otherwise you can translate this string
   freely, literally, or not at all, at your option.
   */
  N_("Auld-lang-syne")

Fix that, and exclude api.scm and card-monkey.scm so they don't have to
be removed manually.

Fixes: https://gitlab.gnome.org/GNOME/aisleriot/-/issues/965