Disable gtk_stock_add
This spews out a million warnings, and was deprecated really early in
the Gtk3 cycle. It is no longer the style to have lots of icons
anyway, so we will get rid of most of these later. For now, disabling
them, but easy to reenable for when we redo the UI.